피드백: 경로 상수화 및 useFetch hook 구현
