+++ src/paths.ts
@@ -3,6 +3,7 @@ export const PATH = {
   LOGIN: "/login",
   MY: "/my",
   ORDER: "/order/:itemId",
+  toORDER: (itemId: string | number) => `/order/${itemId}`,
   THEME: "/themes/:themeId",
   NOTFOUND: "*",
 };
